+++ data/others/test_integration.sh
@@ -25,12 +25,24 @@ echo "Testing Training..."
 
 cd ../trainer/
 
-python Main.py --model_name model_example_1 --load_data imagenet_per90_porp_default
+python Main.py --model_name model_example_1 --load_data imagenet_per90_porp_default --epochs 2
 
 echo "Testing Testing..."
 
 cd ../tester/
 
-python Main.py --model_name model_example_1 --load_data imagenet_per90_porp_default
+python Main.py --model_name model_example_1 --load_data imagenet_per90_porp_default --test_name testing_imagenet_test_set
 
 #python Main.py --model_name model_example_1 --load_data graffiti_per90_porp_default
+
+echo "Testing Analyzer..."
+
+cd ../analyzer/
+
+echo "Testing Analyzer: 1-Top..."
+
+python Main.py --mode top --test_name testing_imagenet_test_set
+
+echo "Testing Analyzer: Deep visualization..."
+
+python Main.py --mode visualization --files ../../data/datasets/graffiti/classes/graffiti/4235365635_a5fba2a2d8_o.jpg --model_name model_example_1 --class_name freight\ car